// frontend/src/components/careerCase/types.ts
// types.ts
export interface CareerCase {
  title: string;
  salary: string;
  experience: string;
  resumeSubmission: string;
  successRate: string;
  employmentDate: string;
  preparationPeriod: string;
  bootcamp: string;
  companyType: string;
  companyLocation: string;
  companySalary: string;
  companySize: string;
  certifications: string;
  techStack: string;
  tips: string[];
  interviewQuestion: string;
}